+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Detect Windows executable and transparently call wine on it
 * Add a loadbranch hotkey to load the entire savestate movie in playback mode
 * Prevent the game screen from appearing unresponsive when the game is paused
+* Don't scroll input editor after manual scrolling, until current frame is not visible
 
 ### Changed